[WFLY-16086] Move tests of wildfly feature pack functionality out of the 'manualmode' testsuite into a new 'manualmode-expansion' testsuite.
Also move the TokenUtil test support class out of testsuite/shared into the only testsuite module that uses it.
