+++ testsuite/integration/manualmode-expansion/src/test/java/org/wildfly/test/manual/management/MPScriptTestCase.java
@@ -35,7 +35,7 @@ import org.junit.runner.RunWith;
 import org.wildfly.core.testrunner.ManagementClient;
 import org.wildfly.core.testrunner.ServerControl;
 import org.wildfly.core.testrunner.ServerController;
-import org.wildfly.core.testrunner.WildflyTestRunner;
+import org.wildfly.core.testrunner.WildFlyRunner;
 
 import jakarta.inject.Inject;
 import org.jboss.as.cli.CommandContext;
@@ -55,7 +55,7 @@ import org.wildfly.core.testrunner.Server;
  *
  * @author jdenise
  */
-@RunWith(WildflyTestRunner.class)
+@RunWith(WildFlyRunner.class)
 @ServerControl(manual = true)
 public class MPScriptTestCase {
