+++ testsuite/integration/manualmode-expansion/src/test/java/org/wildfly/test/manual/microprofile/health/SuccessfulReadinessCheck.java
