+++ testsuite/integration/manualmode-expansion/src/test/java/org/wildfly/test/manual/microprofile/lra/LRAExpressionSupportTestCase.java
@@ -1,15 +1,18 @@
 package org.wildfly.test.manual.microprofile.lra;
 
+import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.as.arquillian.container.ManagementClient;
 import org.jboss.as.controller.PathAddress;
 import org.jboss.as.controller.client.ModelControllerClient;
-import org.jboss.as.test.manualmode.model.AbstractExpressionSupportTestCase;
+import org.jboss.as.test.integration.management.base.AbstractExpressionSupportTestCase;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 
 import java.io.IOException;
 
 import static org.jboss.as.controller.operations.common.Util.createAddOperation;
 
+@RunWith(Arquillian.class)
 public class LRAExpressionSupportTestCase extends AbstractExpressionSupportTestCase {
 
     private static final String JBOSSAS = "jbossas-custom";
